moved more files to standalone and updated routes
