+++ src/interface/src/app/app.component.html
@@ -1,3 +1,5 @@
 <!-- Toolbar -->
 <app-top-bar (toggleEvent)="toggleSidebar($event)"></app-top-bar>
-<app-navigation></app-navigation>
+<section>
+  <router-outlet></router-outlet>
+</section>
