+++ src/interface/src/app/app.component.scss
@@ -6,6 +6,6 @@
   height: 100%;
 }
 
-app-navigation {
+section {
   height: calc(100% - $toolbar-height);
 }
