+++ src/interface/src/app/app.component.spec.ts
@@ -1,12 +1,9 @@
-import { TopBarComponent } from './top-bar/top-bar.component';
-import { NavigationComponent } from './navigation/navigation.component';
 import { of } from 'rxjs';
 import { AuthService } from '@services';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
-import { MockComponent } from 'ng-mocks';
 import { FeaturesModule } from './features/features.module';
 
 describe('AppComponent', () => {
@@ -21,11 +18,7 @@ describe('AppComponent', () => {
     );
     await TestBed.configureTestingModule({
       imports: [HttpClientTestingModule, RouterTestingModule, FeaturesModule],
-      declarations: [
-        AppComponent,
-        MockComponent(NavigationComponent),
-        MockComponent(TopBarComponent),
-      ],
+      declarations: [AppComponent],
       providers: [{ provide: AuthService, useValue: fakeAuthService }],
     }).compileComponents();
