--- src/interface/src/app/navigation/navigation.component.html
@@ -1,1 +1,0 @@
-<router-outlet></router-outlet>
