+++ src/interface/src/app/shared/top-bar/top-bar.component.scss
@@ -1,6 +1,6 @@
-@import '../shared/_variables.scss';
-@import "../../styles/colors";
-@import "../../styles/mixins";
+@import '../../shared/_variables.scss';
+@import "../../../styles/colors";
+@import "../../../styles/mixins";
 
 // Top-bar styling
 .mat-toolbar {
