+++ src/interface/src/app/standalone/delete-dialog/delete-dialog.component.scss
