+++ src/interface/src/app/standalone/delete-dialog/delete-dialog.component.ts
@@ -3,11 +3,15 @@ import {
   MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA,
   MatLegacyDialogRef as MatDialogRef,
 } from '@angular/material/legacy-dialog';
+import { CommonModule } from '@angular/common';
+import { MaterialModule } from '../../material/material.module';
 
 @Component({
   selector: 'app-delete-dialog',
   templateUrl: './delete-dialog.component.html',
   styleUrls: ['./delete-dialog.component.scss'],
+  standalone: true,
+  imports: [CommonModule, MaterialModule],
 })
 export class DeleteDialogComponent {
   constructor(
